Small improvement for non_regression_tests/cleanup.sh script.

+++ non_regression_tests/cleanup.sh
@@ -4,15 +4,14 @@ startdir=`pwd`
 
 
 for i in `ls -1d non_regression_tests/testcases/*`; do
-  cd $i
-  echo ""
-  echo ""
-  echo "*** $i ***"
-  echo ""
-  
-  source settings.sh
-  
-  rm -f *.h5 *.rst jorek_model* rst_*bin* *.tgz ${extra_remote_files}
-  
-  cd $startdir
+  if [ -d $i ]; then
+    cd $i
+    echo "Cleaning up '$i'"
+    
+    source settings.sh
+    
+    rm -f *.h5 *.rst jorek_model* rst_*bin* *.tgz ${extra_remote_files}
+    
+    cd $startdir
+  fi
 done
